frontend/requirements.txt
Flask
flask-pymongo
pandas
google-generativeai
pymongo
python-dotenv
